Finished rewriting tests for CEP-18.
