+++ modules/Odra.toml
@@ -22,3 +22,8 @@ fqn = "access::Ownable"
 [[contracts]]
 fqn = "ownable_2step::Ownable2Step"
 
+[[contracts]]
+fqn = "cep18_token::Cep18"
+
+[[contracts]]
+fqn = "cep18::cep18_client_contract::Cep18ClientContract"
