+++ modules/src/cep18/mod.rs
@@ -1,4 +1,10 @@
+//! # Cep18 Module
+//! A module that allows implementation of CEP-18 token contracts with Odra framework.
+/// Cep18 module errors.
 pub mod errors;
+/// Cep18 module events.
 pub mod events;
+/// Cep18 module utils.
 pub mod utils;
 mod cep18_client_contract;
+mod tests;
