+++ modules/src/cep18/tests/mod.rs
@@ -1,0 +1,3 @@
+mod allowance;
+mod mint_and_burn;
+mod transfer;
